Revert "remove unused gethitssize, revert getmemsize to size_t"
This reverts commit c71d5922a6d6b651575a7f28d74bffbf1e5e1ff5.

+++ Buffer_shared.cpp
@@ -78,7 +78,7 @@ void FacetHistogramBuffer::Reset() {
 }
 
 size_t FacetHistogramBuffer::GetMemSize() const {
-	size_t sum = 0;
+	int sum = 0;
 	sum += nbHitsHistogram.capacity() * sizeof(double);
 	sum += distanceHistogram.capacity() * sizeof(double);
 #ifdef MOLFLOW
