+++ IntersectAABB_shared.cpp
@@ -102,7 +102,7 @@ void AABBNODE::ComputeBB() {
 }
 
 size_t AABBNODE::GetMemSize() {
-	size_t sum = 0;
+	int sum = 0;
 	if (left) {
 		sum += left->GetMemSize();
 	}
